Use int for port instead of unsigned short

+++ gameserver/network/export/server_factory.h
@@ -41,7 +41,7 @@ class ServerFactory
 {
  public:
   static std::unique_ptr<Server> createServer(boost::asio::io_service* io_service,
-                                              unsigned short port,
+                                              int port,
                                               const Server::Callbacks& callbacks);
 };
